+++ .build.ps1
@@ -72,17 +72,17 @@ task Test {
 	# invoke tests, get the output and result
 	$output = Invoke-Build . Demo\.build.ps1 -Result result | Out-String -Width:9999
 
-	assert ($result.AllTasks.Count -eq 111) $result.AllTasks.Count
-	assert ($result.AllErrorCount -eq 24) $result.AllErrorCount
+	assert ($result.AllTasks.Count -eq 114) $result.AllTasks.Count
+	assert ($result.Tasks.Count -eq 28) $result.Tasks.Count
 
-	assert ($result.Tasks.Count -eq 26) $result.Tasks.Count
+	assert ($result.AllErrorCount -eq 24) $result.AllErrorCount
 	assert ($result.ErrorCount -eq 0) $result.AllErrorCount
 
-	assert ($result.WarningCount -ge 1)
 	assert ($result.AllWarningCount -ge 1)
+	assert ($result.WarningCount -ge 1)
 
-	assert ($result.Messages.Count -ge 1)
 	assert ($result.AllMessages.Count -ge 1)
+	assert ($result.Messages.Count -ge 1)
 
 	if ($SkipTestDiff) { return }